requirements.txt
streamlit
torch
torchvision
opencv-python
pillow
numpy
matplotlib

supabase
flask
flask-cors
fpdf
